+++ src/readme/questions/purpose.js
